refactor(backend): extract media repository builders
