+++ backend/src/lib/media-repository-core.ts
@@ -11,15 +11,15 @@ import {
   type MediaRepository,
 } from '@myway/shared';
 import {
-  extractKeyConcepts,
-  extractKeywords,
   findLecture,
-  normalizeText,
-  splitIntoSegments,
-  summarizeByStyle,
-  buildTimelineMarkers,
   now,
 } from '@myway/shared/lms/media/helpers';
+import {
+  applyAudioExtractionCallback,
+  buildAudioExtraction,
+  buildLectureSummaryNote,
+  buildLectureTranscript,
+} from './media-repository-builders';
 import {
   mapExtractionRow,
   mapNoteRow,
@@ -30,29 +30,6 @@ import {
   type PipelineRow,
   type TranscriptRow,
 } from './media-repository-mappers';
-
-function createId(prefix: string): string {
-  return `${prefix}_${crypto.randomUUID().replaceAll('-', '').slice(0, 12)}`;
-}
-
-
-function inferAudioFormat(source: string | undefined): string {
-  if (!source) {
-    return 'pending';
-  }
-
-  const normalized = source.split('?')[0]?.split('#')[0] ?? '';
-  const extension = normalized.includes('.') ? normalized.split('.').pop()?.toLowerCase() : '';
-  if (!extension) {
-    return 'wav';
-  }
-
-  if (['mp3', 'wav', 'm4a', 'aac', 'ogg', 'flac', 'webm'].includes(extension)) {
-    return extension;
-  }
-
-  return 'wav';
-}
 
 export function createMediaRepository(db: D1Database): MediaRepository {
   const repository: MediaRepository = {
@@ -79,37 +56,10 @@ export function createMediaRepository(db: D1Database): MediaRepository {
         return null;
       }
 
-      const fullText = normalizeText(input.text || lecture.content_text || '');
-      if (!fullText) {
+      const transcript = buildLectureTranscript(lecture, userId, input);
+      if (!transcript) {
         return null;
       }
-
-      const fallbackDurationMs = Math.max(lecture.duration_minutes * 60_000, fullText.length * 40, 180_000);
-      const segments = input.segments?.length
-        ? input.segments.map((segment, index) => ({
-            index,
-            start_ms: Math.max(0, Math.round(segment.start_ms)),
-            end_ms: Math.max(Math.round(segment.start_ms), Math.round(segment.end_ms)),
-            text: normalizeText(segment.text),
-          }))
-        : splitIntoSegments(fullText, input.duration_ms ?? fallbackDurationMs);
-      const durationMs =
-        input.duration_ms ??
-        (segments.length > 0 ? segments[segments.length - 1]?.end_ms ?? fallbackDurationMs : fallbackDurationMs);
-      const wordCount = input.word_count ?? fullText.split(/\s+/).filter(Boolean).length;
-      const transcript: LectureTranscript = {
-        id: createId('trs'),
-        lecture_id: lecture.id,
-        user_id: userId,
-        language: input.language ?? 'ko',
-        full_text: fullText,
-        segments,
-        word_count: wordCount,
-        duration_ms: durationMs,
-        stt_provider: input.stt_provider ?? (input.text ? 'text-derived-stt' : 'demo-stt'),
-        stt_model: input.stt_model ?? 'pseudo-stt-v1',
-        created_at: now(),
-      };
 
       await db.prepare(
         'INSERT INTO lecture_transcripts (id, lecture_id, user_id, language, full_text, segments_json, word_count, duration_ms, stt_provider, stt_model, created_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)',
@@ -154,40 +104,7 @@ export function createMediaRepository(db: D1Database): MediaRepository {
       }
 
       const transcript = await repository.getLectureTranscript(lecture.id);
-      const sourceText = transcript?.full_text ?? lecture.content_text;
-      const style = input.style ?? 'brief';
-      const keyConcepts = extractKeyConcepts(sourceText, style === 'detailed' ? 5 : 3);
-      const keywords = extractKeywords(sourceText, 5);
-      const summaryContent = summarizeByStyle(sourceText, style);
-      const timestamps =
-        style === 'timeline'
-          ? buildTimelineMarkers(transcript?.segments ?? splitIntoSegments(sourceText, Math.max(lecture.duration_minutes * 60_000, 180_000)))
-          : null;
-
-      let noteType: LectureNote['note_type'] = 'ai_summary';
-      let titleSuffix = '핵심 요약';
-      if (style === 'detailed') {
-        noteType = 'ai_detailed';
-        titleSuffix = '상세 요약';
-      } else if (style === 'timeline') {
-        noteType = 'ai_timeline';
-        titleSuffix = '타임라인 요약';
-      }
-
-      const note: LectureNote = {
-        id: createId('note'),
-        lecture_id: lecture.id,
-        user_id: userId,
-        note_type: noteType,
-        title: `${lecture.title} - ${titleSuffix}`,
-        content: summaryContent,
-        key_concepts: keyConcepts,
-        keywords,
-        timestamps,
-        language: input.language ?? 'ko',
-        ai_model: 'demo-summary-v1',
-        created_at: now(),
-      };
+      const { note, keyConcepts, keywords } = buildLectureSummaryNote(lecture, userId, input, transcript);
 
       await db.prepare(
         'INSERT INTO lecture_notes (id, lecture_id, user_id, note_type, title, content, key_concepts_json, keywords_json, timestamps_json, language, ai_model, created_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)',
@@ -237,33 +154,7 @@ export function createMediaRepository(db: D1Database): MediaRepository {
       }
 
       const existingTranscript = await repository.getLectureTranscript(lecture.id);
-      const extraction: AudioExtraction = {
-        id: createId('aud'),
-        lecture_id: lecture.id,
-        user_id: userId,
-        source_type: 'video',
-        source_url: input.video_url ?? '/static/media/demo-lecture.mp4',
-        source_video_key: input.video_asset_key,
-        source_video_name: input.source_file_name,
-        source_content_type: input.source_content_type,
-        source_size_bytes: input.source_size_bytes,
-        language: input.language ?? 'ko',
-        requested_stt_provider: input.stt_provider,
-        requested_stt_model: input.stt_model,
-        processing_job_id: null,
-        processing_error: null,
-        audio_url: input.audio_url ?? null,
-        audio_format: input.audio_url ? inferAudioFormat(input.audio_url) : 'pending',
-        audio_duration_ms: Math.max(lecture.duration_minutes * 60_000, 30_000),
-        sample_rate: 16_000,
-        channels: 1,
-        status: input.audio_url ? 'COMPLETED' : 'PROCESSING',
-        transcript_id: existingTranscript?.id ?? null,
-        stt_status: existingTranscript ? 'COMPLETED' : input.audio_url ? 'PROCESSING' : 'PENDING',
-        created_at: now(),
-        processed_at: input.audio_url ? now() : null,
-        updated_at: now(),
-      };
+      const extraction = buildAudioExtraction(lecture, userId, input, existingTranscript);
 
       await db.prepare(
         'INSERT INTO audio_extractions (id, lecture_id, user_id, source_type, source_url, source_video_key, source_video_name, source_content_type, source_size_bytes, language, requested_stt_provider, requested_stt_model, processing_job_id, processing_error, audio_url, audio_format, audio_duration_ms, sample_rate, channels, status, transcript_id, stt_status, created_at, processed_at, updated_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)',
@@ -316,29 +207,7 @@ export function createMediaRepository(db: D1Database): MediaRepository {
         return null;
       }
 
-      const nextStatus = input.status;
-      const resolvedStatus = current.status === 'COMPLETED' && nextStatus === 'PROCESSING' ? current.status : nextStatus;
-      const next: AudioExtraction = {
-        ...current,
-        processing_job_id: input.processing_job_id ?? current.processing_job_id ?? null,
-        processing_error: input.error_message ?? (resolvedStatus === 'FAILED' ? '오디오 추출에 실패했습니다.' : null),
-        audio_url: input.audio_url ?? current.audio_url ?? null,
-        audio_format: input.audio_format ?? (input.audio_url ? inferAudioFormat(input.audio_url) : current.audio_format),
-        audio_duration_ms: input.audio_duration_ms ?? current.audio_duration_ms,
-        sample_rate: input.sample_rate ?? current.sample_rate,
-        channels: input.channels ?? current.channels,
-        status: resolvedStatus,
-        transcript_id: input.transcript_id ?? current.transcript_id,
-        stt_status:
-          input.stt_status ??
-          (resolvedStatus === 'FAILED'
-            ? 'FAILED'
-            : input.transcript_id
-              ? 'COMPLETED'
-              : current.stt_status),
-        processed_at: resolvedStatus === 'COMPLETED' ? current.processed_at ?? now() : current.processed_at ?? null,
-        updated_at: now(),
-      };
+      const next = applyAudioExtractionCallback(current, input);
 
       await db.prepare(
         'UPDATE audio_extractions SET processing_job_id = ?, processing_error = ?, audio_url = ?, audio_format = ?, audio_duration_ms = ?, sample_rate = ?, channels = ?, status = ?, transcript_id = ?, stt_status = ?, processed_at = ?, updated_at = ? WHERE id = ?',
@@ -348,10 +217,10 @@ export function createMediaRepository(db: D1Database): MediaRepository {
           next.processing_error ?? null,
           next.audio_url ?? null,
           next.audio_format,
-        next.audio_duration_ms,
-        next.sample_rate,
-        next.channels,
-        next.status,
+          next.audio_duration_ms,
+          next.sample_rate,
+          next.channels,
+          next.status,
           next.transcript_id ?? null,
           next.stt_status,
           next.processed_at ?? null,
